Test_for_team_AI_step2/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
pgvector
psycopg2-binary
python-dotenv
sentence-transformers
numpy
